fix joi + gestion erreur api

+++ app.to-do-list/to-do-list.controller/list.controller.js
@@ -1,4 +1,5 @@
 const listDatamapper = require("../to-do-list.datamapper/list.datamapper.js");
+const taskDatamapper = require("../to-do-list.datamapper/task.datamapper.js")
 
 const listController = {
   async getAllList (req, res) {
@@ -15,25 +16,37 @@ const listController = {
   async createOneList (req, res) {
         const {name, userId} = req.body;
         const oneList = await listDatamapper.createOneList(name, userId);
-        res.json("liste crée avec succès");
+        res.json(oneList);
     },
 
   async deleteOneList (req, res) {
         const listId = req.params.listId;
-        const taskByListId = await taskDatamapper.deleteTaskByListId(listId);
-        const oneList = await listDatamapper.deleteOneList(listId);
-        res.json(oneList);
+        const listExisted = await listDatamapper.getOneList(listId)
+        if(listExisted.length === 0){
+          res.status(404).json("message: la liste n'existe pas")
+        } else {
+          const taskByListId = await taskDatamapper.deleteTaskByListId(listId);
+          const listDeleted = await listDatamapper.deleteOneList(listId);
+          res.json("message: la liste a été supprimée avec succès");
+        }
+
     },
     
     async modifyOneList (req,res) {
         const listId = req.params.listId;
-        const {name} = req.body;
+        const listExisted = await listDatamapper.getOneList(listId)
+        if(listExisted.length === 0){
+          res.status(404).json("message: la liste n'existe pas")
+        } else {
+          const {name} = req.body;
         let {position} =  req.body
         if (!position){
           position=0
         }
         const updatedList = await listDatamapper.modifyOneList (name, position, listId);
-        res.json(updatedList);
+        res.json("message: la liste a été modifiée avec succès");
+        }
+       
     }
 }
