+++ app.to-do-list/to-do-list.controller/task.controller.js
@@ -16,24 +16,34 @@ const TaskController = {
         const listId = req.params.listId;
         const {name} = req.body;
         const oneTask = await taskDatamapper.createOneTask(name, listId);
-        res.json("message: tache créée avec succès");
+        res.json(oneTask);
     },
 
     async deleteOneTask (req, res) {
         const taskId = req.params.taskId;
-        const oneTask = await taskDatamapper.deleteOneTask(taskId);
-        res.json("message: tache supprimée avec succès");
+        const taskExisted = await taskDatamapper.getOneTask(taskId)
+        if(taskExisted.length === 0){
+          res.status(404).json("message: la tache n'existe pas")
+        } else {
+          const oneTask = await taskDatamapper.deleteOneTask(taskId);
+          res.json("message: la tache a été supprimée avec succès");
+        }
     },
     
     async modifyOneTask (req,res) {
         const taskId = req.params.taskId;
-        const {name} = req.body;
-        let {position} = req.body;
-        if (!position){
-            position=0
-          }
-        const updatedTask = await taskDatamapper.modifyOneTask(name, position, taskId);
-        res.json("tache modifié avec succès");
+        const taskExisted = await taskDatamapper.getOneTask(taskId)
+        if(taskExisted.length === 0){
+          res.status(404).json("message: la tache n'existe pas")
+        } else {
+          const {name} = req.body;
+          let {position} = req.body;
+          if (!position){
+              position=0
+            }
+          const updatedTask = await taskDatamapper.modifyOneTask(name, position, taskId);
+          res.json("tache modifié avec succès");
+        }
     }
 }
