+++ app.wallet/wallet.controller/document.controller.js
@@ -16,20 +16,30 @@ const documentController = {
         const walletId = req.params.walletId;
         const {name, information, file, icon} = req.body;
         const oneDocument = await documentDatamapper.createOneDocument(name, information, file, icon, walletId);
-        res.json("message: Document créée avec succès");
+        res.json(oneDocument);
     },
 
     async deleteOneDocument (req, res) {
         const documentId = req.params.documentId;
-        const oneDocument = await documentDatamapper.deleteOneDocument(documentId);
-        res.json("message: Document supprimé avec succès");
+        const existedDocument = await documentDatamapper.getOneDocument(documentId);
+        if(existedDocument.length === 0){
+            res.status(404).json("message: le document n'existe pas")
+        } else {
+            const oneDocument = await documentDatamapper.deleteOneDocument(documentId);
+            res.json("message: Document supprimé avec succès");
+        }
     },
     
     async modifyOneDocument (req,res) {
         const documentId = req.params.documentId;
+        const existedDocument = await documentDatamapper.getOneDocument(documentId);
+        if(existedDocument.length === 0){
+            res.status(404).json("message: le document n'existe pas")
+        } else {
         const {description, position} = req.body;
-        const updatedDocument = await documentDatamapper.modifyOneDocument(description, position, documentId);
-        res.json("modification réussie");
+        const updatedDocument = await documentDatamapper.modifyOneDocument(name, information, file, icon, documentId);
+        res.json("message: Document modifié avec succès");
+        }
     }
 }
